requirements.txt
langgraph==1.0.4
langchain==1.1.2
langchain-openai==1.1.0
langchain-community==0.4.1
openai==2.9.0
tavily-python==0.7.14
PyGithub==2.8.1
wikipedia==1.4.0
python-dotenv==1.0.0
requests==2.32.5
gitpython==3.1.45
streamlit==1.28.1
